// backend/src/infrastructure/events/types.ts
export interface DomainEvent<T = any> {
  type: string;
  payload: T;
  timestamp: Date;
  correlationId: string;
  userId?: string; // Usuario que generó el evento (opcional)
}